pyproject.toml
[tool.poetry]
name = "sandbox_rl"
version = "0.0.0"
description = "Sandbox project for experimenting with RL."
license = "MIT"
authors = ["sindre0830 <sindre0830@gmail.com>"]
readme = "README.md"
repository = "https://github.com/sindre0830/template-python"
packages = [{include = "sandbox_rl", from = "src"}]

[tool.poetry.scripts]
main = "sandbox_rl.main:main"

[tool.poetry.dependencies]
python = "^3.12"
python-interface = "^1.6.1"
numpy = "^2.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-custom-exit-code = "^0.3.0"
ipykernel = "^6.29.5"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
